- Added Tenant Id to Tables: Settings, Custodies, Resources, Operations - Updated the api SProc for saving them - Code cleaning

+++ BSharp/dbo/Stored Procedures/dbo.api_Agents__Save.sql
@@ -2,10 +2,11 @@
 	@Agents [AgentList] READONLY
 AS
 BEGIN
-	DECLARE @IdMappings IdMappingList;
+	DECLARE @IdMappings IdMappingList, @TenantId int;
 	BEGIN TRANSACTION
 		BEGIN TRY
-			DELETE FROM dbo.Custodies WHERE Id IN (SELECT Id FROM @Agents WHERE Status = N'Deleted');
+			SELECT @TenantId = dbo.fn_TenantId();
+			DELETE FROM dbo.Custodies WHERE TenantId = @TenantId AND Id IN (SELECT Id FROM @Agents WHERE Status = N'Deleted');
 
 			INSERT INTO @IdMappings([NewId], [OldId])
 			SELECT x.[NewId], x.[OldId]
@@ -13,17 +14,17 @@ BEGIN
 			(
 				MERGE INTO dbo.Custodies AS t
 				USING (
-					SELECT [Id], [AgentType], [Name], [IsActive] 
+					SELECT @TenantId As [TenantId], [Id], [AgentType], [Name], [IsActive] 
 					FROM @Agents 
 					WHERE [Status] IN (N'Inserted', N'Updated')
-				) AS s ON t.Id = s.Id
+				) AS s ON t.[TenantId] = s.[TenantId] AND t.Id = s.Id
 				WHEN MATCHED THEN
 					UPDATE SET 
 						t.[Name] = s.[Name],
 						t.[IsActive] = s.[IsActive]
 				WHEN NOT MATCHED THEN
-					INSERT ([CustodyType], [Name], [IsActive])
-					VALUES (s.[AgentType], s.[Name], s.[IsActive])
+					INSERT ([TenantId], [CustodyType], [Name], [IsActive])
+					VALUES (@TenantId, s.[AgentType], s.[Name], s.[IsActive])
 				--WHEN NOT MATCHED BY SOURCE THEN 
 				--	DELETE
 				OUTPUT inserted.[Id] As [NewId], s.[Id] As [OldId]
@@ -31,10 +32,10 @@ BEGIN
 
 			MERGE INTO dbo.Agents t
 			USING (
-				SELECT M.[NewId] As [Id], [AgentType], [IsRelated] , [UserId], [TaxIdentificationNumber], [RegisteredAddress], [Title], [Gender], [BirthDateTime]
+				SELECT @TenantId As [TenantId], M.[NewId] As [Id], [AgentType], [IsRelated] , [UserId], [TaxIdentificationNumber], [RegisteredAddress], [Title], [Gender], [BirthDateTime]
 				FROM @Agents I 
 				JOIN @IdMappings M ON I.Id = M.OldId
-			) As s ON t.Id = s.Id
+			) AS s ON t.[TenantId] = s.[TenantId] AND t.Id = s.Id
 			WHEN MATCHED THEN
 				UPDATE SET
 					t.[Id]						= s.[Id],
@@ -46,8 +47,8 @@ BEGIN
 					t.[Gender]					= s.[Gender],
 					t.[BirthDateTime]			= s.[BirthDateTime]
 			WHEN NOT MATCHED THEN
-				INSERT ([Id], [AgentType]	,	[IsRelated] ,	[UserId],	[TaxIdentificationNumber],	[RegisteredAddress], [Title], [Gender], [BirthDateTime])
-				VALUES (s.[Id], s.[AgentType], s.[IsRelated], s.[UserId], s.[TaxIdentificationNumber], s.[RegisteredAddress], s.[Title], s.[Gender], s.[BirthDateTime]);
+				INSERT ([TenantId], [Id], [AgentType]	,	[IsRelated] ,	[UserId],	[TaxIdentificationNumber],	[RegisteredAddress], [Title], [Gender], [BirthDateTime])
+				VALUES (@TenantId, s.[Id], s.[AgentType], s.[IsRelated], s.[UserId], s.[TaxIdentificationNumber], s.[RegisteredAddress], s.[Title], s.[Gender], s.[BirthDateTime]);
 
 				--WHEN NOT MATCHED BY SOURCE THEN 
 				--	DELETE
@@ -70,7 +71,8 @@ BEGIN
     COMMIT TRANSACTION; 
 	
 	SELECT C.[Id], A.[AgentType], C.[Name], C.[IsActive], A.[IsRelated], A.[UserId], A.[TaxIdentificationNumber], A.[RegisteredAddress], A.[Title], A.[Gender], A.[BirthDateTime], N'Unchanged' As [Status], M.[OldId] As [TemporaryId]
-	FROM dbo.Custodies C JOIN dbo.Agents A ON C.Id = A.Id
+	FROM dbo.Custodies C 
+	JOIN dbo.Agents A ON C.[TenantId] = A.[TenantId] AND C.Id = A.Id  
 	LEFT JOIN @IdMappings M ON C.[Id] = M.[NewId]
 	WHERE C.[Id] IN (
 		SELECT M.[NewId] FROM @Agents A JOIN @IdMappings M ON A.Id = M.OldId WHERE [Status] = N'Inserted'
