+++ BSharp/dbo/Stored Procedures/dbo.api_Resources__Save.sql
@@ -2,11 +2,11 @@
 	@Resources [ResourceList] READONLY
 AS
 BEGIN
-	DECLARE @IdMappings IdMappingList;
+	DECLARE @IdMappings IdMappingList, @TenantId int;
 	BEGIN TRANSACTION
 		BEGIN TRY
-
-			DELETE FROM dbo.Resources WHERE Id IN (SELECT Id FROM @Resources WHERE Status = N'Deleted');
+			SELECT @TenantId = dbo.fn_TenantId();
+			DELETE FROM dbo.Resources WHERE TenantId = @TenantId AND Id IN (SELECT Id FROM @Resources WHERE Status = N'Deleted');
 
 			INSERT INTO @IdMappings([NewId], [OldId])
 			SELECT x.[NewId], x.[OldId]
@@ -14,10 +14,10 @@ BEGIN
 			(
 				MERGE INTO dbo.Resources AS t
 				USING (
-					SELECT [Id], [ResourceType], [Name], [Code], [UnitOfMeasure], [Memo], [Lookup1], [Lookup2], [Lookup3], [Lookup4], [GoodForServiceParentId], [FungibleParentId]
+					SELECT @TenantId As [TenantId], [Id], [ResourceType], [Name], [Code], [UnitOfMeasure], [Memo], [Lookup1], [Lookup2], [Lookup3], [Lookup4], [GoodForServiceParentId], [FungibleParentId]
 					FROM @Resources 
 					WHERE [Status] IN (N'Inserted', N'Updated')
-				) AS s ON t.Id = s.Id
+				) AS s ON t.[TenantId] = s.[TenantId] AND t.Id = s.Id
 				WHEN MATCHED THEN
 					UPDATE SET 
 						t.[ResourceType]			= s.[ResourceType],         
@@ -32,8 +32,8 @@ BEGIN
 						t.[GoodForServiceParentId]	= s.[GoodForServiceParentId],
 						t.[FungibleParentId]		= s.[FungibleParentId]
 				WHEN NOT MATCHED THEN
-					INSERT ([ResourceType],		[Name],		 [Code], [UnitOfMeasure],	 [Memo],	[Lookup1],	[Lookup2],	[Lookup3],		[Lookup4], [GoodForServiceParentId], [FungibleParentId])
-					VALUES (s.[ResourceType], s.[Name], s.[Code], s.[UnitOfMeasure], s.[Memo], s.[Lookup1], s.[Lookup2], s.[Lookup3], s.[Lookup4], s.[GoodForServiceParentId], s.[FungibleParentId])
+					INSERT ([TenantId], [ResourceType],		[Name],	[Code], [UnitOfMeasure],	 [Memo],	[Lookup1],	[Lookup2],	[Lookup3],		[Lookup4], [GoodForServiceParentId], [FungibleParentId])
+					VALUES (@TenantId, s.[ResourceType], s.[Name], s.[Code], s.[UnitOfMeasure], s.[Memo], s.[Lookup1], s.[Lookup2], s.[Lookup3], s.[Lookup4], s.[GoodForServiceParentId], s.[FungibleParentId])
 				--WHEN NOT MATCHED BY SOURCE THEN 
 				--	DELETE
 				OUTPUT inserted.[Id] As [NewId], s.[Id] As [OldId]
@@ -68,7 +68,8 @@ BEGIN
 	SELECT R.[Id], R.[ResourceType], R.[Name], R.[Code], R.[UnitOfMeasure], R.[Memo], R.[Lookup1], R.[Lookup2], R.[Lookup3], R.[Lookup4], R.[GoodForServiceParentId], R.[FungibleParentId], N'Unchanged' As [Status], M.[OldId] As [TemporaryId]
 	FROM dbo.Resources R
 	LEFT JOIN @IdMappings M ON R.[Id] = M.[NewId]
-	WHERE R.[Id] IN (
+	WHERE R.[TenantId] = @TenantId
+	AND R.[Id] IN (
 		SELECT M.[NewId] FROM @Resources R JOIN @IdMappings M ON R.Id = M.OldId WHERE [Status] = N'Inserted'
 		UNION ALL
 		SELECT Id FROM @Resources WHERE [Status] = N'Updated'
